+++ modules/http-proxy/src/upstream/lb/hash_ring.rs
@@ -136,6 +136,7 @@ mod tests {
             priority: 0,
             connection_timeout: None,
             idle_timeout: std::time::Duration::from_secs(60),
+            dns_status: Default::default(),
         })
     }
 
@@ -149,6 +150,7 @@ mod tests {
             priority: 0,
             connection_timeout: None,
             idle_timeout: std::time::Duration::from_secs(60),
+            dns_status: Default::default(),
         })
     }